extracting annot
