+++ include/vrv/editorial.h
@@ -150,41 +150,6 @@ private:
 };
 
 //----------------------------------------------------------------------------
-// Annot
-//----------------------------------------------------------------------------
-
-class Annot : public EditorialElement, public TextListInterface, public AttPlist, public AttSource {
-public:
-    /**
-     * @name Constructors, destructors, and other standard methods
-     * Reset method reset all attribute classes
-     */
-    ///@{
-    Annot();
-    virtual ~Annot();
-    virtual void Reset();
-    virtual std::string GetClassName() const { return "Annot"; }
-    virtual ClassId GetClassId() const { return ANNOT; }
-    ///@}
-
-    /**
-     * Add a text element to an annotation.
-     * Only supported elements will be actually added to the child list.
-     */
-    virtual void AddChild(Object *object);
-
-private:
-    //
-public:
-    /**
-     * A copy of the annot content tree stored as pugi::xml_document
-     */
-    pugi::xml_document m_content;
-
-private:
-};
-
-//----------------------------------------------------------------------------
 // App
 //----------------------------------------------------------------------------
